% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg.annotate.R
\name{cpg.annotate}
\alias{cpg.annotate}
\title{cpg.annotate}
\usage{
cpg.annotate(
  datatype = c("array", "sequencing"),
  object,
  what = c("Beta", "M"),
  arraytype = c("EPIC", "450K", "mouse"),
  analysis.type = c("differential", "variability", "ANOVA", "diffVar"),
  design,
  contrasts = FALSE,
  cont.matrix = NULL,
  fdr = 0.05,
  coef,
  varFitcoef = NULL,
  topVarcoef = NULL,
  ...
)
}
\arguments{
\item{datatype}{Character string representing the type of data being analysed (array or sequencing).}

\item{object}{Either:
\itemize{
\item A matrix of M-values, with unique Illumina probe IDs as rownames and unique sample IDs as column names or,
\item A GenomicRatioSet, appropriately annotated.
}}

\item{what}{Does the data matrix contain Beta or M-values? Not needed if object is a GenomicRatioSet.}

\item{arraytype}{Is the data matrix sourced from mouse, EPIC or 450K data?}

\item{analysis.type}{"differential" for dmrcate() to return DMRs; "variability" to return VMRs; "ANOVA" to return "whole experiment" DMRs, incorporating all possible contrasts from the design matrix using the moderated F-statistics; "diffVar" to return differentially variable methylated regions, using the missMethyl package to generate t-statistics.}

\item{design}{Study design matrix. Identical context to differential analysis pipeline in limma. Must have an intercept if contrasts=FALSE. Applies only when analysis.type \%in\% c("differential", "ANOVA", "diffVar").}

\item{contrasts}{Logical denoting whether a limma-style contrast matrix is specified. Only applicable when datatype="array" and analysis.type \%in\% c("differential", "diffVar").}

\item{cont.matrix}{Limma-style contrast matrix for explicit contrasting. For each call to cpg.annotate, only one contrast will be fit. Only applicable when datatype="array" and analysis.type \%in\% c("differential", "diffVar").}

\item{fdr}{FDR cutoff (Benjamini-Hochberg) for which CpG sites are individually called as significant. Used to index default thresholding in dmrcate(). Highly recommended as the primary thresholding parameter for calling DMRs. Not used when analysis.type == "variability".}

\item{coef}{The column index in design corresponding to the phenotype comparison. Corresponds to the comparison of interest in design when contrasts=FALSE, otherwise must be a column name in cont.matrix. Only applicable when analysis.type == "differential".}

\item{varFitcoef}{The columns of the design matrix containing the comparisons to test for differential variability. If left NULL, will test all columns. Identical context to missMethyl::varFit(). Only applicable when analysis.type \%in\% "diffVar".}

\item{topVarcoef}{Column number or column name specifying which coefficient of the linear model fit is of interest. It should be the same coefficient that the differential variability testing was performed on. Default is last column of fit object. Identical context to missMethyl::topVar(). Only applicable when analysis.type \%in\% "diffVar".}

\item{...}{Extra arguments passed to the limma function lmFit() (analysis.type="differential").}
}
\value{
Annotated CpG object
}
\description{
Helper function so DMRcate will be compatible with the mouse manifest. Corresponds to DMRcate::cpg.annotate.
}
\references{
Ritchie, M. E., Phipson, B., Wu, D., Hu, Y., Law, C. W., Shi, W., & Smyth, G. K. (2015). limma powers differential expression analyses for RNA-sequencing and microarray studies. Nucleic Acids Research, 43(7), e47.

Feng, H., Conneely, K. N., & Wu, H. (2014). A Bayesian hierarchical model to detect differentially methylated loci from single nucleotide resolution sequencing data. Nucleic Acids Research, 42(8), e69.

Phipson, B., & Oshlack, A. (2014). DiffVar: a new method for detecting differential variability with application to methylation in cancer and aging. Genome Biol, 15(9), 465.

Peters T.J., Buckley M.J., Statham, A., Pidsley R., Samaras K., Lord R.V., Clark S.J. and Molloy P.L. De novo identification of differentially methylated regions in the human genome. Epigenetics & Chromatin 2015, 8:6, doi:10.1186/1756-8935-8-6.
}
\author{
Tim J. Peters \href{mailto:t.peters@garvan.org.au}{t.peters@garvan.org.au}, Chiara Herzog \href{mailto:chiara.herzog@uibk.ac.at}{chiara.herzog@uibk.ac.at} (modifications)
}
